Add/Change: (define-widget ement-room-membership) Event as value
The Widget library is still mysterious to me, but some trial-and-error
helps.  This makes the ement-room-membership widget's value the Matrix
event struct, and it uses a :value-create function to insert the
string.

This pattern can probably be used for other widgets in the future,
e.g. for event type-specific widgets.