Change: Save session struct
This seems to work correctly, but it wouldn't surprise me if some
unreadable types accidentally end up in the data and cause a read
error at some point (which this tries to account for).  If so, those
types should be omitted from the saved session struct.

Also, this effectively saves all room events locally, which means they
will accumulate over time.  This is probably undesirable, and some
kind of limiting or filtering should be implemented.