Change/Fix: (ement--write-session) Clear unnecessary slots
I found that some ement-user structs were not being read properly (and
who knows why--tracing that down may be tedious), which caused errors
when rooms were displayed after loading the session, so for now we'll
just clear these slots.  This means that the first sync will be an
initial sync, so we aren't persisting the whole session.  That may not
be ideal, but it's much simpler and avoids a whole lot of potential
bugs, too.