Add: (ement-session-events) Track seen events in hash table
We already have all of these event structs in the rooms' event lists,
so this is just extra pointers to the same data, and extra copies of
the event IDs, which seems reasonable.  We need to track this, for now
at least, because it lets us ensure that we aren't reusing transaction
IDs on a session (next commit).