WIP: Compare chunk size against batch-size
This seems like the most reliable way to determine whether we are
finished filling (if the chunk is smaller than the batch-size, there
must be no more events).  It seems like comparing the tokens ought to
work, but it doesn't seem to (maybe something in the spec I don't
grok).