Fix malformed m.receipt events causing error in process sentinel
Sometimes I get this error: `error in process sentinel: Wrong type
argument: listp, "{\"ts\": 1553242544352}"` when syncing. It's
happening in the m.receipt event handler. It seems like for some
reason this event, but not all events, hasn't been decoded from json,
or needs to be decoded a second time.

Hopefully this patch is minimal and harmless. An alternative behavior
would be to drop the malformed receipt events.