Fix: (room-defevent "m.room.message") Replaced messages
When reopening closed room buffers that contain retro-loaded messages
that include events that have been replaced by later edits, the
original messages were inserted alongside the edits, because the edit
events were inserted before the original events, so the original
events were not replaced by the edits.

This happened because the edits were processed before the originals,
because the events were stored out-of-order in the timeline.  Now,
this shouldn't have been the case, because when retro-loading
messages, we store them at the beginning of the timeline, before the
newer events.  But, for some reason, this isn't always working.  The
Matrix spec says that the definition of "chronological order" is
up to the server, so maybe it's not always returning them in the order
we would expect.

So this is a mildly hacky workaround: if the event's unsigned data
indicates it's been replaced, we just don't insert it.  Theoretically,
we should already have any such replacing event.

If we wanted to be extra careful, we could look up the replacing event
in the hash table to ensure we've got it, but we can add that later if
necessary; for now, let's not.