Fix: (ement-room-mark-read) Marking latest event
The last event in a room's timeline may not always be the last event
displayed in the room's buffer (e.g. a reaction to an earlier event).
So when intending to mark all events in the room as read (when point
is after the last displayed event), mark the latest event as read
rather than the last displayed event.  (Otherwise, the room would not
get fully marked as read, and could display as an unread room.)