Fix: (ement-room--m.image-callback) Workaround rare bug
I'm not sure what can cause this.  While I've heard a couple of
reports from users, this is the first time it's happened to me.  It
might be caused by an earlier error interrupting sync processing,
leaving a room data or buffer in an inconsistent state.  Anyway, this
should prevent it from being a problem, while still leaving a way to
potentially learn the cause so it can be truly solved.