Change: (ement-taxy-room-avatar-cache) Make buffer-local
On the rare occasion that the ement-taxy buffer is opened before room
avatars have been fetched, the generated avatars would fill the cache,
and the downloaded ones would never be shown.  Now, killing the buffer
will discard the cache (which is a good idea, anyway, otherwise it
becomes almost like a memory leak).