Change/Fix: Room list avatars
Since the cached avatar needs to be discarded when a room avatar event
is received, it's simpler to store the cached avatar in the room's
local alist rather than use a buffer-local hash table in the room list
buffer.  It does mean that the avatars will remain in memory when the
room list buffer is killed, but that's unlikely to cause significant
memory usage, and the cached avatars could be cleared easily enough if
needed.

This also fixes a bug or race condition in that generated avatars
would not be replaced with downloaded ones.