Change: (ement-with-room-and-session) Don't use pcase/map plist form
It would be preferable to use the form (map :prompt-form), but that
causes, e.g. Emacs 27.2 on Debian 11 to produce
macro-expansion/void-variable errors on installation when the version
of `map' is too old.  Since there's no way to force Emacs to actually
load the newer version after installing it, there's no way to prevent
users from suffering the broken installation and having to reinstall
this package to fix it.  So we'll avoid using that nicer form for some
number of years, until older versions of Emacs or `map' are not widely
used.