Add/Change: (ement--json-parse-buffer) Use Jansson functions
The json.el docstrings and the JSON section of the Elisp info manual
don't make this clear, but in order to use the libjansson functions
for parsing JSON, one must call them; json-read, et al remain
Elisp-based.

Some simple testing shows that this results in a 3-5x speed increase
for parsing JSON responses; for a 38 MB response, the numbers show
23.6 seconds using json-read, and 4.1 seconds using the Jansson-based
json-parse-buffer.

Thanks to Ryan Rix (@rrix) for discovering this!