Change/Fix: (ement--json-parse-buffer) Fall back to json-read
A user reported that JSON-PARSE-BUFFER signaled this error while
parsing a sync response:

  json-parse-error "too big integer near '18446744073407678578'"

I don't know why such a large integer (larger than
most-positive-fixnum on my system) would be in a JSON object returned
by the matrix.org server (the user's homeserver), but I confirmed that
JSON-READ can parse it without error, so we'll fall back to it.

And temporarily we'll have this message when it happens, so I can try
to gauge how often it's happening.