Change: (ement-room-init-compose-buffer): Set yank-excluded-properties
Add `line-prefix' and `wrap-prefix' to `yank-excluded-properties' in
compose buffers.

Depending on `ement-room-message-format-spec', messages may include
these text properties for the purposes of formatting in the room
buffer, in which case copying such text and yanking it into another
buffer results in these properties also being applied.

As this is a common (and undesirable) scenario for compose buffers,
we cater to it by default.