Fix: (ement-room-image-scale) Prevent scrolling issues with images
When maximising an image, eliminate all image padding so that the line
height will not exceed the window body height (or width), and scroll
the window to the top of the image to ensure it is displayed in full.

Also take the horizontal indentation of the image into account when
setting the maximum width, so that the image does not extend beyond
the right-hand edge of the window.