Use a window-parameter for the compose buffer window auto-height cache
Previously this was a buffer-local variable, which was not correct,
but was also fine most of the time (as we only care strongly about the
size of the selected window, and we flush the cache when switching
between windows, and in most cases a compose buffer will only ever be
displayed in a single window).

This commit also eliminates the timer-triggered resize from
`ement-room-compose-buffer-window-buffer-change-handler' as it no
longer seems necessary, so `post-command-hook' is now the sole trigger
for resizing.  (The call to `window-preserve-size' added in commit
7418626abe79fce12f834251d4690dcdf4331dd7 is likely another factor in
being able to remove this code, although that alone had not been
sufficient.)