Change/Fix: Cache computed displaynames
When a user's displayname was unset, we did not cache the computed
displayname.  In most cases that seemed to be no problem; but in the
case of a direct room in which the other user had no displayname set,
it would mean iterating over the events and looking for one setting
the displayname every time that EMENT--USER-DISPLAYNAME-IN is called.

When the room list buffer is updated, that function is called for each
such room; over time, as such rooms' gain more events, that means
taking longer to iterate over them, which means taking longer to
update the room list buffer.

Now we cache the computed displayname even if unset; and if a new
member event comes in that sets the displayname, we force
recalculation.

Fixes #298.

Reported-by: Rutherther <https://github.com/Rutherther>